changed InputFile and OutputFile with FileUtils APIs
