+++ extended/src/main/java/apoc/export/parquet/ExportParquetFileStrategy.java
@@ -3,10 +3,10 @@ package apoc.export.parquet;
 import apoc.Pools;
 import apoc.export.util.ProgressReporter;
 import apoc.result.ProgressInfo;
+import apoc.util.FileUtils;
 import apoc.util.QueueBasedSpliterator;
 import apoc.util.QueueUtil;
 import apoc.util.Util;
-import org.apache.hadoop.fs.Path;
 import org.apache.parquet.hadoop.ParquetWriter;
 import org.apache.parquet.hadoop.example.ExampleParquetWriter;
 import org.apache.parquet.schema.MessageType;
@@ -22,8 +22,6 @@ import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
-
-import static apoc.export.parquet.ParquetUtil.getParquetFile;
 
 
 public abstract class ExportParquetFileStrategy<TYPE, IN> implements ExportParquetStrategy<IN, Stream<ProgressInfo>> {
@@ -51,13 +49,12 @@ public abstract class ExportParquetFileStrategy<TYPE, IN> implements ExportParqu
         progressInfo.batchSize = config.getBatchSize();
         ProgressReporter reporter = new ProgressReporter(null, null, progressInfo);
 
-        String parquetFile = getParquetFile(fileName);
-        Path fileToWrite = new Path(parquetFile);
         final BlockingQueue<ProgressInfo> queue = new ArrayBlockingQueue<>(10);
         Util.inTxFuture(pools.getDefaultExecutorService(), db, tx -> {
             int batchCount = 0;
             List<TYPE> rows = new ArrayList<>(config.getBatchSize());
-            ExampleParquetWriter.Builder builder = ExampleParquetWriter.builder(fileToWrite);
+            ParquetBufferedWriter parquetBufferedWriter = new ParquetBufferedWriter(FileUtils.getOutputStream(fileName));
+            ExampleParquetWriter.Builder builder = ExampleParquetWriter.builder(parquetBufferedWriter);
 
             try {
                 Iterator<TYPE> it = toIterator(reporter, data);
